LPS-119966 Applying classNameId filter only to DDM structures
The filter by classNameId is there to retrieve the subtypes (structures)
of JournalArticle. Other entities (like FileEntries) don't have structures
associated and use other entities as subtype (DLFileEntryType). This is
why filtering by classNameId returns no DLFileEntryTypes because they
are not associated to the main entity classNameId.

Now the query does this:

ClassName = DDMStructure and classNameId = X
OR
ClassName == FileEntry